+++ lib/Completion/Bridge/WorseReflection/Completor/ContextSensitiveCompletor.php
@@ -7,6 +7,7 @@ use Microsoft\PhpParser\Node;
 use Microsoft\PhpParser\Node\DelimitedList\ArgumentExpressionList;
 use Microsoft\PhpParser\Node\Expression\ArgumentExpression;
 use Microsoft\PhpParser\Node\Expression\CallExpression;
+use Microsoft\PhpParser\Node\Expression\ObjectCreationExpression;
 use Microsoft\PhpParser\Node\QualifiedName;
 use Phpactor\Completion\Bridge\TolerantParser\Qualifier\AlwaysQualfifier;
 use Phpactor\Completion\Bridge\TolerantParser\TolerantCompletor;
@@ -14,8 +15,12 @@ use Phpactor\Completion\Bridge\TolerantParser\TolerantQualifiable;
 use Phpactor\Completion\Bridge\TolerantParser\TolerantQualifier;
 use Phpactor\TextDocument\ByteOffset;
 use Phpactor\TextDocument\TextDocument;
-use Phpactor\WorseReflection\Bridge\TolerantParser\Reflection\AbstractReflectionMethodCall;
 use Phpactor\WorseReflection\Core\Exception\NotFound;
+use Phpactor\WorseReflection\Core\Inference\Context\ClassLikeContext;
+use Phpactor\WorseReflection\Core\Inference\Context\MemberAccessContext;
+use Phpactor\WorseReflection\Core\Reflection\Collection\ReflectionParameterCollection;
+use Phpactor\WorseReflection\Core\Reflection\ReflectionFunctionLike;
+use Phpactor\WorseReflection\Core\Reflection\ReflectionMember;
 use Phpactor\WorseReflection\Core\Type\ArrayType;
 use Phpactor\WorseReflection\Core\Type\ClassLikeType;
 use Phpactor\WorseReflection\Core\Util\NodeUtil;
@@ -71,9 +76,9 @@ class ContextSensitiveCompletor implements TolerantCompletor, TolerantQualifiabl
     private function resolveFilterableType(Node $node, TextDocument $source, ByteOffset $offset): ?ClassLikeType
     {
         $argumentNb = 0;
-        $callExpression = $node;
+        $memberAccessOrObjectCreation = $node;
         if ($node instanceof QualifiedName) {
-            $callExpression = null;
+            $memberAccessOrObjectCreation = null;
             $argumentExpression = $node->getFirstAncestor(ArgumentExpression::class);
             if ($argumentExpression instanceof ArgumentExpression) {
                 $list = $argumentExpression->getFirstAncestor(ArgumentExpressionList::class);
@@ -81,33 +86,73 @@ class ContextSensitiveCompletor implements TolerantCompletor, TolerantQualifiabl
                     return null;
                 }
                 $argumentNb = NodeUtil::argumentOffset($list, $argumentExpression) ?? 0;
-                $callExpression = $list->parent;
+                $memberAccessOrObjectCreation = $list->parent;
             }
         }
         if ($node instanceof ArgumentExpressionList) {
             $argumentNb = count(iterator_to_array($node->getValues()));
-            $callExpression = $node->parent;
+            $memberAccessOrObjectCreation = $node->parent;
         }
 
-        if (!$callExpression instanceof CallExpression) {
+        if (!$memberAccessOrObjectCreation instanceof CallExpression && !$memberAccessOrObjectCreation instanceof ObjectCreationExpression) {
             return null;
         }
 
+        $offset = $memberAccessOrObjectCreation->openParen?->getStartPosition();
+        if (null === $offset) {
+            return null;
+        }
         try {
-            $callExpression = $this->reflector->reflectNode($source, $callExpression->openParen->getStartPosition());
+            ;
+            $memberAccessOrObjectCreation = $this->reflector->reflectOffset($source, $offset)->nodeContext();
         } catch (NotFound $e) {
             return null;
         }
-        if (!$callExpression instanceof AbstractReflectionMethodCall) {
-            return null;
+        if ($memberAccessOrObjectCreation instanceof MemberAccessContext) {
+            return $this->typeFromMemberAccess($memberAccessOrObjectCreation, $argumentNb);
         }
 
+        if ($memberAccessOrObjectCreation instanceof ClassLikeContext) {
+            return $this->typeFromClassInstantiation($memberAccessOrObjectCreation, $argumentNb);
+        }
+
+
+        return null;
+    }
+    /**
+     * @param MemberAccessContext<ReflectionMember> $memberAccessOrObjectCreation
+     */
+    private function typeFromMemberAccess(MemberAccessContext $memberAccessOrObjectCreation, int $argumentNb): ?ClassLikeType
+    {
         try {
-            $functionLike = $callExpression->method();
-            $parameters = $functionLike->parameters();
+            $functionLike = $memberAccessOrObjectCreation->accessedMember();
         } catch (NotFound) {
             return null;
         }
+        if (!$functionLike instanceof ReflectionFunctionLike) {
+            return null;
+        }
+        $parameters = $functionLike->parameters();
+        return $this->typeFromParameters($parameters, $argumentNb);
+    }
+
+    /**
+     * @param int<0, max> $argumentNb
+     */
+    private function typeFromClassInstantiation(ClassLikeContext $classLikeContext, int $argumentNb): ?ClassLikeType
+    {
+        try {
+            $classLike = $classLikeContext->classLike();
+            $constructor = $classLike->methods()->get('__construct');
+            $parameters = $constructor->parameters();
+        } catch (NotFound) {
+            return null;
+        }
+        return $this->typeFromParameters($parameters, $argumentNb);
+    }
+
+    private function typeFromParameters(ReflectionParameterCollection $parameters, int $argumentNb): ?ClassLikeType
+    {
         $parameter = $parameters->at($argumentNb);
         if (null === $parameter) {
             return null;