+++ lib/Indexer/Model/Record/HasPath.php
@@ -2,12 +2,17 @@
 
 namespace Phpactor\Indexer\Model\Record;
 
+use Phpactor\TextDocument\TextDocumentUri;
+
 interface HasPath
 {
     /**
      * @return $this
      */
-    public function setFilePath(string $filePath);
+    public function setFilePath(TextDocumentUri $filePath);
 
+    /**
+     * Rename to URI
+     */
     public function filePath(): ?string;
 }