+++ lib/Indexer/Tests/Adapter/IndexBuilderTestCase.php
@@ -49,7 +49,7 @@ abstract class IndexBuilderTestCase extends IntegrationTestCase
             'IamAnAttribute',
             function (ClassRecord $record): void {
                 self::assertInstanceOf(ClassRecord::class, $record);
-                self::assertEquals($this->workspace()->path('project/attribute.php'), $record->filePath());
+                self::assertEquals('file://' . $this->workspace()->path('project/attribute.php'), $record->filePath());
                 self::assertEquals('IamAnAttribute', $record->fqn());
                 self::assertEquals(38, $record->start()->toInt());
                 self::assertEquals(52, $record->end()->toInt());
@@ -62,7 +62,7 @@ abstract class IndexBuilderTestCase extends IntegrationTestCase
             'ThisClass',
             function (ClassRecord $record): void {
                 self::assertInstanceOf(ClassRecord::class, $record);
-                self::assertEquals($this->workspace()->path('project/test.php'), $record->filePath());
+                self::assertEquals('file://' . $this->workspace()->path('project/test.php'), $record->filePath());
                 self::assertEquals('ThisClass', $record->fqn());
                 self::assertEquals(12, $record->start()->toInt());
                 self::assertEquals(21, $record->end()->toInt());
@@ -163,7 +163,7 @@ abstract class IndexBuilderTestCase extends IntegrationTestCase
             'ThisInterface',
             function (ClassRecord $record): void {
                 self::assertInstanceOf(ClassRecord::class, $record);
-                self::assertEquals($this->workspace()->path('project/test.php'), $record->filePath());
+                self::assertEquals('file://' . $this->workspace()->path('project/test.php'), $record->filePath());
                 self::assertEquals('ThisInterface', $record->fqn());
                 self::assertEquals(16, $record->start()->toInt());
                 self::assertEquals(29, $record->end()->toInt());
@@ -184,7 +184,7 @@ abstract class IndexBuilderTestCase extends IntegrationTestCase
             'ThisTrait',
             function (ClassRecord $record): void {
                 self::assertInstanceOf(ClassRecord::class, $record);
-                self::assertEquals($this->workspace()->path('project/test.php'), $record->filePath());
+                self::assertEquals('file://' . $this->workspace()->path('project/test.php'), $record->filePath());
                 self::assertEquals('ThisTrait', $record->fqn());
                 self::assertEquals(12, $record->start()->toInt());
                 self::assertEquals(21, $record->end()->toInt());
@@ -221,7 +221,7 @@ abstract class IndexBuilderTestCase extends IntegrationTestCase
             'SomeEnum',
             function (ClassRecord $record): void {
                 self::assertInstanceOf(ClassRecord::class, $record);
-                self::assertEquals($this->workspace()->path('project/test.php'), $record->filePath());
+                self::assertEquals('file://' . $this->workspace()->path('project/test.php'), $record->filePath());
                 self::assertEquals('SomeEnum', $record->fqn());
                 self::assertEquals(11, $record->start()->toInt());
                 self::assertEquals(19, $record->end()->toInt());
@@ -433,7 +433,7 @@ abstract class IndexBuilderTestCase extends IntegrationTestCase
             , 'Barfoos\foobar',
             function (FunctionRecord $record): void {
                 self::assertCount(1, $record->references());
-                self::assertEquals($this->workspace()->path('project/test1.php'), $record->filePath());
+                self::assertEquals('file://' . $this->workspace()->path('project/test1.php'), $record->filePath());
             }
         ];
     }